Add useContext to manage state of todos
