+++ src/components/TodosLogic.jsx
@@ -1,75 +1,13 @@
-import { useState, useEffect } from 'react';
-import { v4 as uuidv4 } from 'uuid';
+import { TodosProvider } from '../context/TodosContext';
 
 import InputTodo from './InputTodo';
 import TodosList from './TodosList';
 
-const TodosLogic = () => {
-  function getInitialTodos() {
-    const temp = localStorage.getItem('todos');
-    const savedTodos = JSON.parse(temp);
-    return savedTodos || [];
-  }
-
-  const [todos, setTodos] = useState(getInitialTodos());
-
-  const handleChange = (id) => {
-    setTodos((prevState) => prevState.map((todo) => {
-      if (todo.id === id) {
-        return {
-          ...todo,
-          completed: !todo.completed,
-        };
-      }
-      return todo;
-    }));
-  };
-
-  const delTodo = (id) => {
-    setTodos([
-      ...todos.filter((todo) => (todo.id !== id)),
-    ]);
-  };
-
-  const addTodoItem = (title) => {
-    const newTodo = {
-      id: uuidv4(),
-      title,
-      completed: false,
-    };
-    setTodos([...todos, newTodo]);
-  };
-
-  const setUpdate = (updatedTitle, id) => {
-    setTodos(
-      todos.map((todo) => {
-        if (todo.id === id) {
-          // eslint-disable-next-line no-param-reassign
-          todo.title = updatedTitle;
-        }
-        return todo;
-      }),
-    );
-  };
-
-  useEffect(() => {
-    const temp = JSON.stringify(todos);
-    localStorage.setItem('todos', temp);
-  }, [todos]);
-
-  return (
-    <div>
-      <InputTodo
-        addTodoItem={addTodoItem}
-      />
-      <TodosList
-        todosProps={todos}
-        handleChange={handleChange}
-        delTodo={delTodo}
-        setUpdate={setUpdate}
-      />
-    </div>
-  );
-};
+const TodosLogic = () => (
+  <TodosProvider>
+    <InputTodo />
+    <TodosList />
+  </TodosProvider>
+);
 
 export default TodosLogic;
